Add full support of ADD_I5/6/7/8

+++ sources/ecudata.h
@@ -228,8 +228,8 @@ typedef struct ecudata_t
 
  uint16_t op_comp_code;                  //!< Contains code of operation for packet being sent - OP_COMP_NC packet
  uint16_t op_actn_code;                  //!< Contains code of operation for packet being received - OP_COMP_NC packet
- uint16_t ecuerrors_for_transfer;        //!< Buffering of error codes being sent via UART in real time
- uint16_t ecuerrors_saved_transfer;      //!< Buffering of error codes for read/write from/to EEPROM which is being sent/received
+ uint32_t ecuerrors_for_transfer;        //!< Buffering of error codes being sent via UART in real time
+ uint32_t ecuerrors_saved_transfer;      //!< Buffering of error codes for read/write from/to EEPROM which is being sent/received
  uint8_t  use_knock_channel_prev;        //!< Previous state of knock channel's usage flag
 
  uint8_t engine_mode;                    //!< Current engine mode(start, idle, work)
